refactored tableSidebar
