+++ src/components/index.ts
@@ -2,7 +2,7 @@ export { Profile } from './profile';
 export { SignUp, Login, Popup, ConfirmPassword, CreateNewPassword } from './auth';
 export { Navigation } from './navigation';
 export { Preloader } from './preloader';
-export { TableSidebar } from './tableSidebar';
+export { TableSidebar } from './packsManagement';
 export { CustomButton, CustomInput, Modal } from './customComponents';
 export { PageError404 } from './404';
 export { Cards, DeleteCardModal } from './cards';
