+++ src/components/packsManagement/PacksList.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
 import { Table } from '../table';
-import { TableSidebar } from '../tableSidebar';
 
 import stylePacks from './PacksManagement.module.css';
+import { TableSidebar } from './tableSidebar';
 
 export const PacksList = (): any => (
   <div className={stylePacks.container}>
