+++ src/components/packsManagement/PacksManagement.module.css
@@ -3,7 +3,6 @@
     width: 100%;
 }
 .container {
-    position: absolute;
     width: 1008px;
     height: 500px;
     margin: 84px 136px;
@@ -11,4 +10,5 @@
     border-radius: 8px;
     display: flex;
     flex-direction: row;
+    justify-content: center;
 }
