+++ src/components/packsManagement/index.ts
@@ -1,0 +1,1 @@
+export { TableSidebar } from './tableSidebar';
