+++ src/components/packsManagement/tableSidebar/index.ts
@@ -1,1 +1,2 @@
 export { TableSidebar } from './TableSidebar';
+export { Range } from './Range';
