+++ src/components/packsManagement/tableSidebar/tableSidebar.module.css
@@ -1,0 +1,36 @@
+.tableSidebar {
+    min-width: 243px;
+    background-color: #ECECF9;
+    border-radius: inherit;
+}
+
+.buttonsContainer {
+    margin: 24px 0;
+    width: 100%;
+    height: 36px;
+    display: flex;
+    justify-content: center;
+    border-radius: 2px;
+}
+
+.button {
+    width: 100px;
+    text-align: center;
+    border: 1px solid #25142E;
+}
+
+.header3 {
+    margin: 24px;
+    font-size: 1.1em;
+}
+
+/*.pickBlock {
+    padding: 1em;
+}*/
+
+@media (max-width: 1250px) {
+    .tableSidebar {
+        flex-direction: row;
+        width: 100%;
+    }
+}
