+++ src/components/table/index.ts
@@ -1,2 +1,1 @@
 export { Table } from './Table';
-export { SuperRange } from './Range';
