--- src/components/tableSidebar/tableSidebar.module.css
@@ -1,27 +1,0 @@
-.tableSidebar {
-    width: 243px;
-    background-color: #ECECF9;
-    border-radius: 8px 0 0 8px;
-}
-
-.buttonsContainer {
-    height: 36px;
-    display: flex;
-    justify-content: space-evenly;
-}
-
-.header3 {
-    margin: 33px 87px 20px 24px;
-    font-size: 1.1em;
-}
-
-/*.pickBlock {
-    padding: 1em;
-}*/
-
-@media (max-width: 1250px){
-    .tableSidebar{
-        flex-direction: row;
-        width: 100%;
-    }
-}
